:sparkles: [#1325] New JSON action in admin form builder
