+++ src/openforms/js/components/admin/form_design/logic/actions/Actions.js
@@ -10,6 +10,8 @@ import {ComponentsContext} from '../../../forms/Context';
 import StepSelection from '../StepSelection';
 import {Action as ActionType, ActionError} from './types';
 import DSLEditorNode from '../DSLEditorNode';
+import {FeatureFlagsContext, FormVariablesContext} from '../../Context';
+import JsonWidget from '../../../forms/JsonWidget';
 
 const ActionProperty = ({action, errors, onChange}) => {
   const modifiablePropertyChoices = Object.entries(MODIFIABLE_PROPERTIES).map(([key, info]) => [
@@ -133,6 +135,32 @@ const ActionValue = ({action, errors, onChange}) => {
   );
 };
 
+const ActionVariableValue = ({action, errors, onChange}) => {
+  const allVariables = useContext(FormVariablesContext);
+
+  const getVariableChoices = variables => {
+    return variables.map(variable => [variable.key, variable.name]);
+  };
+
+  return (
+    <>
+      <DSLEditorNode errors={errors.variable}>
+        {/*TODO: This should be a searchable select for when there are a billion variables?*/}
+        <Select
+          name="variable"
+          choices={getVariableChoices(allVariables)}
+          allowBlank
+          onChange={onChange}
+          value={action.variable}
+        />
+      </DSLEditorNode>
+      <DSLEditorNode errors={errors.action?.value}>
+        <JsonWidget name="action.value" logic={action.action.value} onChange={onChange} />
+      </DSLEditorNode>
+    </>
+  );
+};
+
 const ActionStepNotApplicable = ({action, errors, onChange}) => {
   return (
     <DSLEditorNode errors={errors.formStep}>
@@ -150,6 +178,10 @@ const ActionComponent = ({action, errors, onChange}) => {
     }
     case 'value': {
       Component = ActionValue;
+      break;
+    }
+    case 'variable': {
+      Component = ActionVariableValue;
       break;
     }
     case '':
